+++ src/js/util/Templates.js
